requirements.txt
 asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.9.2
gunicorn==20.0.4
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.4.1
